ai-module/requirements.txt
flask
python-dotenv
google-generativeai
flask-cors
duckduckgo-search
beautifulsoup4
requests
googlesearch-python
